git-svn-id: file:///Users/billburke/jboss/resteasy/resteasy-git/svn-server-sync/resteasy/trunk/rest-star/messaging@1091 2b1ed4c4-5db3-0410-90e4-80a7a6204c25

+++ session/src/main/java/org/jboss/resteasy/star/messaging/queue/DestinationSettings.java
@@ -8,7 +8,7 @@ public class DestinationSettings
 {
    protected boolean duplicatesAllowed;
    private boolean durableSend;
-   private long consumerSessionTimeoutSeconds = 10;
+   private int consumerSessionTimeoutSeconds = 10;
 
    public boolean isDuplicatesAllowed()
    {
@@ -20,12 +20,12 @@ public class DestinationSettings
       this.duplicatesAllowed = duplicatesAllowed;
    }
 
-   public long getConsumerSessionTimeoutSeconds()
+   public int getConsumerSessionTimeoutSeconds()
    {
       return consumerSessionTimeoutSeconds;
    }
 
-   public void setConsumerSessionTimeoutSeconds(long consumerSessionTimeoutSeconds)
+   public void setConsumerSessionTimeoutSeconds(int consumerSessionTimeoutSeconds)
    {
       this.consumerSessionTimeoutSeconds = consumerSessionTimeoutSeconds;
    }
